backend/requirements.txt
asgiref==3.9.1
black==24.8.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
cryptography==45.0.7
defusedxml==0.7.1
Django==4.2
django-filter==25.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.0
drf-extra-fields==3.7.0
filetype==1.2.0
flake8==7.1.1
gunicorn==22.0.0
idna==3.10
isort==5.13.2
mccabe==0.7.0
mypy_extensions==1.1.0
oauthlib==3.3.1
packaging==25.0
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.4.0
psycopg2-binary==2.9.9
pycodestyle==2.12.1
pycparser==2.23
pyflakes==3.2.0
PyJWT==2.10.1
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2025.2
reportlab==4.4.4
requests==2.32.5
requests-oauthlib==2.0.0
ruff==0.13.0
social-auth-app-django==5.4.3
social-auth-core==4.7.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.15.0
urllib3==2.5.0
